patch: try to fix cast error for float

+++ src/main/kotlin/online/danielstefani/paddy/stats/dto/AveragePowerEveryTemporal.kt
@@ -1,7 +1,7 @@
 package online.danielstefani.paddy.stats.dto
 
 class AveragePowerEveryTemporal {
-    var averageW: Float? = null
+    var averageW: Double? = null
 
     var temporal: Long? = null
 }
